Update: test

+++ models/model_zoo/pidnet/model_utils.py
@@ -17,7 +17,7 @@ def segmentation_head(x_in, interplanes, outplanes, scale_factor=None):
 
     x = layers.BatchNormalization(momentum=bn_mom)(x)
     x = layers.Activation("relu")(x)
-    x = layers.Conv2D(outplanes, kernel_size=(1, 1), use_bias=range, padding="valid", dtype=tf.float32)(x)  # bias difference
+    x = layers.Conv2D(outplanes, kernel_size=(1, 1), use_bias=range, padding="valid")(x)  # bias difference
 
     if scale_factor is not None:
         input_shape = tf.keras.backend.int_shape(x)
